user promt to begin
At this point, the progra requires some dummy input to enter the while loop. We can fix this with a do while loop, boolean flags, or many other ways. When we decide to fix this problem, we need to be careful because there are many if, else, and other conditional statements contained in this big while loop; thus making alterations could mess things up. So for now, we prompt the user to enter some dummy input in a nice way "type begin to search". Of course, the input does not have to be "begin", it just has to be anything other than "exit". When the dummy input is provided and is not "exit" the program continues. Again, this problem can be corrected in a number of ways and doing so would have to be done carefully.